// src/main/webapp/src/api/models/DistanceMatrix.ts
/* tslint:disable */
/* eslint-disable */
/**
 * dicoop API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.13.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DistanceMatrix
 */
export interface DistanceMatrix {
    /**
     * 
     * @type {Array<string>}
     * @memberof DistanceMatrix
     */
    locations?: Array<string>;
    /**
     * 
     * @type {Array<Array<number>>}
     * @memberof DistanceMatrix
     */
    distances?: Array<Array<number>>;
}

/**
 * Check if a given object implements the DistanceMatrix interface.
 */
export function instanceOfDistanceMatrix(value: object): value is DistanceMatrix {
    return true;
}

export function DistanceMatrixFromJSON(json: any): DistanceMatrix {
    return DistanceMatrixFromJSONTyped(json, false);
}

export function DistanceMatrixFromJSONTyped(json: any, ignoreDiscriminator: boolean): DistanceMatrix {
    if (json == null) {
        return json;
    }
    return {
        
        'locations': json['locations'] == null ? undefined : json['locations'],
        'distances': json['distances'] == null ? undefined : json['distances'],
    };
}

export function DistanceMatrixToJSON(json: any): DistanceMatrix {
    return DistanceMatrixToJSONTyped(json, false);
}

export function DistanceMatrixToJSONTyped(value?: DistanceMatrix | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'locations': value['locations'],
        'distances': value['distances'],
    };
}